src/requirements.txt
atproto==0.0.56
Pillow==10.2.0
python-telegram-bot==20.7
mongoengine==0.27.0
motor==3.4.0
pymongo==4.8.0
